refactor: add auth to getServerConstants
