+++ .gitignore
@@ -37,3 +37,4 @@ playwright/temp
 # OpenAPI error logs
 openapi-ts-error-*.log
 .output
+tmp/
